Move ApiKeyService to Jellyfin.Api
